Fade out animation across all projects complete 1. Take the current project and raise its z-index to 1 2. Load the new project with z-index of 0, opacity of 1 3. The current project's opacity drops from 1 to 0 at 0% to 100% 4. The z-index of the current project drops from 1 to -1 at 99% to 100% Repeat for every project swap.
Had to get rid of flexbox for centering the projects and
instead use absolute positioning and centering with a
transform since flexbox doesn't seem to handle z-indexing
too well (or at all?).

The only issue with this is:
all loaded projects affect the flow of the document since they
have zero opacity but are still displaying as a block. Unfortunately,
display can not be animated and doing it through js became too
complicated since - recall - editing innerHTML of an element means
we lose reference to all variables that pointed to children of that
element.

However, I'd argue that for this project its okay!
Since we are "swapping" between projects, they should really all
be the same dimmensions within the browser anyway. At least they
should be, cause I havent started resizing for different browser
sizes yet, haha.